+++ test/unit/MMM-RBB-Weather-Test.js
@@ -1,7 +1,7 @@
 const assert = require('assert');
 const sinon = require('sinon');
 
-const newModule = require('./ModuleTestEnv').newModule;
+const { newModule } = require('./ModuleTestEnv');
 
 describe('MMM-RBB-Weather', () => {
 
@@ -168,6 +168,55 @@ describe('MMM-RBB-Weather', () => {
 
             // Assert
             assert.deepStrictEqual(data.module, module);
+        });
+    });
+
+    describe('getCurrentText', () => {
+
+        it('should return same text if splitting is disabled', () => {
+
+            // Arrange
+            module.config.splitCurrentTextGreater = 0;
+
+            // Act
+            let text = module.getCurrentText('too small');
+
+            // Assert
+            assert.strictEqual(text, 'too small');
+        });
+
+        it('should return same text if it is too small to split', () => {
+
+            // Act
+            let text = module.getCurrentText('too small');
+
+            // Assert
+            assert.strictEqual(text, 'too small');
+        });
+
+        it('should return same text if it is long enough but has no colon', () => {
+
+            // Arrange
+            let text = 'this text is long enough but has no colon so it is not splitted';
+
+            // Act
+            let notSplittedText = module.getCurrentText(text);
+
+            // Assert
+            assert.strictEqual(notSplittedText, text);
+        });
+
+        it('should return splitted text if it is long enough', () => {
+
+            // Arrange
+            let text = 'this text is long enough and has a colon, so it is splitted there';
+            let expected = 'this text is long enough and has a colon<br/>so it is splitted there';
+
+            // Act
+            let splittedText = module.getCurrentText(text);
+
+            // Assert
+            assert.strictEqual(splittedText, expected);
         });
     });